# memory pools for walkers