```
benchmark                       old ns/op     new ns/op     delta
BenchmarkApplyNewObject-12      1363050       1347076       -1.17%
BenchmarkUpdateNewObject-12     1228947       1160335       -5.58%
BenchmarkRepeatedUpdate-12      396604        384544        -3.04%
BenchmarkSetToFields-12         5107          5214          +2.10%
BenchmarkFieldsToSet-12         10170         10118         -0.51%
BenchmarkYAMLToTyped-12         54439         51352         -5.67%

benchmark                       old allocs     new allocs     delta
BenchmarkApplyNewObject-12      4214           4089           -2.97%
BenchmarkUpdateNewObject-12     3661           3568           -2.54%
BenchmarkRepeatedUpdate-12      1179           1111           -5.77%
BenchmarkSetToFields-12         14             14             +0.00%
BenchmarkFieldsToSet-12         82             82             +0.00%
BenchmarkYAMLToTyped-12         90             74             -17.78%
```